Fix cosmetic dark mode issue
On devices with a notch (iPhones for sure) the area around the notch
was light themed, despite the dark mode setting.

This fix makes this area match the background color of the site,
like in the light mode.

JavaScript must bo in control of this, to make the dynamic change
possible. Base.html is changed, because the area blinked from time to time
(it was rendered light and changed dynamically miliseconds later, it was visible).
I wanted it to match the expected initial value, to prevent that.